+++ src/main/java/insane96mcp/survivalreimagined/event/SREventFactory.java
@@ -2,6 +2,9 @@ package insane96mcp.survivalreimagined.event;
 
 import insane96mcp.survivalreimagined.module.misc.level.SRExplosion;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.MinecraftForge;
 
 public class SREventFactory {
@@ -16,6 +19,12 @@ public class SREventFactory {
     public static void onSRExplosionCreated(SRExplosion explosion)
     {
         SRExplosionCreatedEvent event = new SRExplosionCreatedEvent(explosion);
+        MinecraftForge.EVENT_BUS.post(event);
+    }
+
+    public static void onBlockBurnt(Level level, BlockPos pos, BlockState state)
+    {
+        BlockBurntEvent event = new BlockBurntEvent(level, pos, state);
         MinecraftForge.EVENT_BUS.post(event);
     }
 }